+++ css-inline/src/lib.rs
@@ -33,33 +33,15 @@ mod html;
 mod parser;
 
 pub use error::InlineError;
-use html5ever::tendril::StrTendril;
 use indexmap::IndexMap;
-use smallvec::{smallvec, SmallVec};
 use std::{
     borrow::Cow,
-    collections::btree_map::Entry,
     io::{ErrorKind, Write},
 };
 
 use hasher::BuildNoHashHasher;
 use html::{Document, Specificity};
 pub use url::{ParseError, Url};
-
-/// Replace double quotes in property values.
-///
-/// This implementation is deliberately simplistic and covers only `font-family`, but escaping
-/// might be needed in other properties that accept strings.
-macro_rules! replace_double_quotes {
-    ($target:expr, $name:expr, $value:expr) => {
-        // Avoid allocation if there is no double quote in the input string
-        if $name.starts_with("font-family") && memchr::memchr(b'"', $value.as_bytes()).is_some() {
-            $target.push_str(&$value.replace('"', "\'"))
-        } else {
-            $target.push_str($value)
-        };
-    };
-}
 
 /// Configuration options for CSS inlining process.
 #[derive(Debug)]
@@ -227,18 +209,7 @@ impl<'a> CSSInliner<'a> {
     ///   - Internal CSS selector parsing error;
     #[inline]
     pub fn inline_to<W: Write>(&self, html: &str, target: &mut W) -> Result<()> {
-        let document = self.inline_impl(html)?;
-        document.serialize(
-            target,
-            self.options.keep_style_tags,
-            self.options.keep_link_tags,
-        )?;
-        Ok(())
-    }
-
-    /// Non-generic inlining function.
-    fn inline_impl(&self, html: &str) -> Result<Document> {
-        let mut document =
+        let document =
             Document::parse_with_options(html.as_bytes(), self.options.preallocate_node_capacity);
         // CSS rules may overlap, and the final set of rules applied to an element depend on
         // selectors' specificity - selectors with higher specificity have more priority.
@@ -316,39 +287,13 @@ impl<'a> CSSInliner<'a> {
                 // Which means that they will fall into this category and will be ignored
             }
         }
-        let mut style_buffer: SmallVec<[String; 8]> = smallvec![];
-        for (node_id, styles) in styles.iter_mut() {
-            let element = if let Some(element) = document[*node_id].as_not_ignored_element_mut() {
-                element
-            } else {
-                continue;
-            };
-            styles.sort_unstable_by(|_, (a, _), _, (b, _)| a.cmp(b));
-            match element.attributes.get_style_entry() {
-                Entry::Vacant(entry) => {
-                    let estimated_declaration_size: usize = styles
-                        .iter()
-                        .map(|(name, (_, value))| {
-                            name.len()
-                                .saturating_add(STYLE_SEPARATOR.len())
-                                .saturating_add(value.len())
-                                // Additional byte for the semicolon symbol
-                                .saturating_add(1)
-                        })
-                        .sum();
-                    let mut buffer = String::with_capacity(estimated_declaration_size);
-                    for (property, (_, value)) in styles {
-                        write_declaration(&mut buffer, property, value);
-                        buffer.push(';');
-                    }
-                    entry.insert(buffer.into());
-                }
-                Entry::Occupied(mut entry) => {
-                    merge_styles(entry.get_mut(), styles, &mut style_buffer)?;
-                }
-            }
-        }
-        Ok(document)
+        document.serialize(
+            target,
+            styles,
+            self.options.keep_style_tags,
+            self.options.keep_link_tags,
+        )?;
+        Ok(())
     }
 
     fn get_full_url<'u>(&self, href: &'u str) -> Cow<'u, str> {
@@ -440,138 +385,4 @@ pub fn inline(html: &str) -> Result<String> {
 #[inline]
 pub fn inline_to<W: Write>(html: &str, target: &mut W) -> Result<()> {
     CSSInliner::default().inline_to(html, target)
-}
-
-const STYLE_SEPARATOR: &str = ": ";
-
-macro_rules! push_or_update {
-    ($style_buffer:expr, $length:expr, $name: expr, $value:expr) => {
-        if let Some(style) = $style_buffer.get_mut($length) {
-            style.clear();
-            style.push_str($name);
-            style.push_str(STYLE_SEPARATOR);
-            style.push_str($value.trim());
-        } else {
-            let value = $value.trim();
-            let mut style = String::with_capacity(
-                $name
-                    .len()
-                    .saturating_add(STYLE_SEPARATOR.len())
-                    .saturating_add(value.len()),
-            );
-            style.push_str($name);
-            style.push_str(STYLE_SEPARATOR);
-            style.push_str(value);
-            $style_buffer.push(style);
-            $length = $length.saturating_add(1);
-        }
-    };
-}
-
-#[inline]
-fn write_declaration(style: &mut String, name: &str, value: &str) {
-    style.push_str(name);
-    style.push_str(STYLE_SEPARATOR);
-    replace_double_quotes!(style, name, value.trim());
-}
-
-/// Merge a new set of styles into an current one, considering the rules of CSS precedence.
-///
-/// The merge process maintains the order of specificity and respects the `!important` rule in CSS.
-fn merge_styles(
-    current_style: &mut StrTendril,
-    new_styles: &IndexMap<&str, (Specificity, &str)>,
-    declarations_buffer: &mut SmallVec<[String; 8]>,
-) -> Result<()> {
-    // This function is designed with a focus on reusing existing allocations where possible
-    // We start by parsing the current declarations in the "style" attribute
-    let mut parser_input = cssparser::ParserInput::new(current_style);
-    let mut parser = cssparser::Parser::new(&mut parser_input);
-    let current_declarations =
-        cssparser::DeclarationListParser::new(&mut parser, parser::CSSDeclarationListParser);
-    // We manually manage the length of our buffer. The buffer may contain slots used
-    // in previous runs, and we want to access only the portion that we build in this iteration
-    let mut parsed_declarations_count: usize = 0;
-    for (idx, declaration) in current_declarations.enumerate() {
-        parsed_declarations_count = parsed_declarations_count.saturating_add(1);
-        let (property, value) = declaration?;
-        let estimated_declaration_size = property
-            .len()
-            .saturating_add(STYLE_SEPARATOR.len())
-            .saturating_add(value.len());
-        // We store the existing style declarations in the buffer for later merging with new styles
-        // If possible, we reuse existing slots in the buffer to avoid additional allocations
-        if let Some(buffer) = declarations_buffer.get_mut(idx) {
-            buffer.clear();
-            buffer.reserve(estimated_declaration_size);
-            write_declaration(buffer, &property, value);
-        } else {
-            let mut buffer = String::with_capacity(estimated_declaration_size);
-            write_declaration(&mut buffer, &property, value);
-            declarations_buffer.push(buffer);
-        };
-    }
-    // Next, we iterate over the new styles and merge them into our existing set
-    // New rules will not override old ones unless they are marked as `!important`
-    for (property, (_, value)) in new_styles {
-        match (
-            value.strip_suffix("!important"),
-            declarations_buffer.iter_mut().find(|style| {
-                style.starts_with(property)
-                    && style.get(property.len()..=property.len().saturating_add(1))
-                        == Some(STYLE_SEPARATOR)
-            }),
-        ) {
-            // The new rule is `!important` and there's an existing rule with the same name
-            // In this case, we override the existing rule with the new one
-            (Some(value), Some(buffer)) => {
-                // We keep the rule name and the colon-space suffix - '<rule>: `
-                buffer.truncate(property.len().saturating_add(STYLE_SEPARATOR.len()));
-                buffer.push_str(value.trim());
-            }
-            // There's no existing rule with the same name, but the new rule is `!important`
-            // In this case, we add the new rule with the `!important` suffix removed
-            (Some(value), None) => {
-                push_or_update!(
-                    declarations_buffer,
-                    parsed_declarations_count,
-                    property,
-                    value
-                );
-            }
-            // There's no existing rule with the same name, and the new rule is not `!important`
-            // In this case, we just add the new rule as-is
-            (None, None) => push_or_update!(
-                declarations_buffer,
-                parsed_declarations_count,
-                property,
-                value
-            ),
-            // Rule exists and the new one is not `!important` - leave the existing rule as-is and
-            // ignore the new one.
-            (None, Some(_)) => {}
-        }
-    }
-
-    // We can now dispose of the parser input, which allows us to clear the current style string
-    drop(parser_input);
-    // Now we prepare to write the merged styles back into current style
-    current_style.clear();
-    let size_estimate: usize = declarations_buffer[..parsed_declarations_count]
-        .iter()
-        // Additional byte for the semicolon symbol
-        .map(|s| s.len().saturating_add(1))
-        .sum();
-
-    // Reserve enough space in current style for the merged style string
-    current_style.reserve(u32::try_from(size_estimate).expect("Size overflow"));
-
-    // Write the merged styles into the current style
-    for declaration in &declarations_buffer[..parsed_declarations_count] {
-        if !current_style.is_empty() {
-            current_style.push_char(';');
-        }
-        current_style.push_slice(declaration);
-    }
-    Ok(())
 }